requirements.txt
scapy>=2.4.5
PyYAML>=6.0
SQLAlchemy>=1.4.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
Flask>=2.0.0
python-iptables>=1.0.0 